refactor: 태그 저장 책임 수정 (#4)
 - Command가 아닌 UseCase로 네이밍 변경

- 태그 조회 및 저장이 한 트랜잭션 안에서 이루어지도록 로직 수정
