+++ RevUp-api/src/main/java/com/revup/question/usecase/CreateQuestionUseCase.java
@@ -1,4 +1,4 @@
-package com.revup.question.command;
+package com.revup.question.usecase;
 
 import com.revup.image.entity.QuestionImage;
 import com.revup.question.dto.request.QuestionCreateRequest;
@@ -6,7 +6,6 @@ import com.revup.question.dto.response.QuestionIdResponse;
 import com.revup.question.entity.Question;
 import com.revup.question.mapper.QuestionMapper;
 import com.revup.question.service.QuestionService;
-import com.revup.tag.adaptor.TagAdaptor;
 import com.revup.tag.entity.Tag;
 import com.revup.tag.mapper.TagMapper;
 import com.revup.user.adaptor.UserAdaptor;
@@ -18,12 +17,12 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class QuestionCommand {
+public class CreateQuestionUseCase {
     private final QuestionService questionService;
     private final QuestionMapper questionMapper;
     private final TagMapper tagMapper;
     private final UserAdaptor userAdaptor;
-    private final TagAdaptor tagAdaptor;
+
 
     public QuestionIdResponse create(QuestionCreateRequest request) {
 //        User user = userAdaptor.getCurrentUser();
@@ -31,15 +30,12 @@ public class QuestionCommand {
 
         Question question = questionMapper.toEntity(request, user);
 
-        List<Tag> tags = tagAdaptor.getTags(tagMapper.toString(request.tags()));
 
         List<QuestionImage> questionImages = questionMapper.toQuestionImages(request.images(), question);
 
-        Long id = questionService.createQuestion(question, tags, questionImages);
+        Long id = questionService.createQuestion(question, tagMapper.toNameList(request.tags()), questionImages);
 
         return new QuestionIdResponse(id);
     }
-
-
 
 }
